+++ dali/test/python/test_functional_api.py
@@ -114,3 +114,26 @@ def test_scalar_constant():
     assert(np.array_equal(arr1, ref1))
     assert(np.array_equal(arr2, ref2))
     assert(np.array_equal(arr3, ref3))
+
+def test_to_snake_case_impl():
+    fn_name_tests = [
+        ('Test', 'test'),
+        ('OneTwo', 'one_two'),
+        ('TestXYZ', 'test_xyz'),
+        ('testX', 'test_x'),
+        ('TestXx', 'test_xx'),
+        ('testXX', 'test_xx'),
+        ('OneXYZTwo', 'one_xyz_two'),
+        ('MFCC', 'mfcc'),
+        ('RandomBBoxCrop', 'random_bbox_crop'),
+        ('STFT_CPU', 'stft_cpu'),
+        ('DOUBLE__UNDERSCORE', 'double__underscore'),
+        ('double__underscore', 'double__underscore'),
+        ('XYZ1ABC', 'xyz1abc'),
+        ('XYZ1abc', 'xyz1abc'),
+        ('trailing__', 'trailing__'),
+        ('TRAILING__', 'trailing__')
+    ]
+
+    for inp, out in fn_name_tests:
+        assert fn._to_snake_case(inp) == out, f"{fn._to_snake_case(inp)} != {out}"
